Switching over to Jade file management.
Switching HTML to Jade format for easier manipulation
Reworked the folder structure a bit so that the server is in the root
folder. This makes it much easier to access content.
Added a function to the DatabaseWrapper - getAllSections. Allows us to
get the names of every seciton in the database.